Interactive and Responsive Video Player
I was given a video file and a general site mockup design. Using the
Media Element jquery plugin, jquery, javascript, and css I was able to
publish my own unique video player. The Video player was then formatted
to be usable on all media types and screen sizes. Lastly, using the
flex box display the rest of the site was formatted to be responsive as
well.